# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LogVision — Apache access log analytics API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LogUploadInput {
  label: string;
}

export interface LogSession {
  id: number;
  filename: string;
  label: string;
  totalLines: number;
  parsedLines: number;
  /** @nullable */
  dateFrom: string | null;
  /** @nullable */
  dateTo: string | null;
  createdAt: string;
}

export interface AppSummary {
  name: string;
  requestCount: number;
  uniqueIPs: number;
  peakHour: number;
  peakDayOfWeek: number;
  /** @nullable */
  topPage?: string | null;
}

export interface HeatmapCell {
  dayOfWeek: number;
  hour: number;
  requestCount: number;
}

export interface HourlyTotal {
  hour: number;
  requestCount: number;
}

export interface AppPatterns {
  appName: string;
  heatmap: HeatmapCell[];
  hourlyTotals: HourlyTotal[];
}

export type PageStatStatusCodes = { [key: string]: number };

export interface PageStat {
  url: string;
  requestCount: number;
  uniqueIPs: number;
  /** @nullable */
  avgBytes?: number | null;
  statusCodes: PageStatStatusCodes;
  hourlyDistribution?: HourlyTotal[];
}

export interface OverviewStats {
  totalRequests: number;
  uniqueIPs: number;
  totalApps: number;
  totalPages: number;
  /** @nullable */
  dateFrom: string | null;
  /** @nullable */
  dateTo: string | null;
  avgRequestsPerHour: number;
  peakHour: number;
  peakDay: number;
}

export interface TimelinePoint {
  timestamp: string;
  requestCount: number;
  /** @nullable */
  uniqueIPs?: number | null;
}

export interface StatusCodeStat {
  statusCode: number;
  count: number;
  percentage: number;
}

export type ListAppsParams = {
  /**
   * @nullable
   */
  sessionId?: number | null;
};

export type GetAppPatternsParams = {
  /**
   * @nullable
   */
  sessionId?: number | null;
  /**
   * @nullable
   */
  dateFrom?: string | null;
  /**
   * @nullable
   */
  dateTo?: string | null;
};

export type GetAppPagesParams = {
  /**
   * @nullable
   */
  sessionId?: number | null;
  /**
   * @nullable
   */
  hourFrom?: number | null;
  /**
   * @nullable
   */
  hourTo?: number | null;
  /**
   * @nullable
   */
  limit?: number | null;
};

export type GetOverviewParams = {
  /**
   * @nullable
   */
  sessionId?: number | null;
};

export type GetHeatmapParams = {
  /**
   * @nullable
   */
  sessionId?: number | null;
  /**
   * @nullable
   */
  appName?: string | null;
};

export type GetTimelineParams = {
  /**
   * @nullable
   */
  sessionId?: number | null;
  /**
   * @nullable
   */
  appName?: string | null;
  /**
   * @nullable
   */
  granularity?: GetTimelineGranularity;
};

export type GetTimelineGranularity =
  | (typeof GetTimelineGranularity)[keyof typeof GetTimelineGranularity]
  | null;

export const GetTimelineGranularity = {
  hour: "hour",
  day: "day",
} as const;

export type GetStatusCodesParams = {
  /**
   * @nullable
   */
  sessionId?: number | null;
  /**
   * @nullable
   */
  appName?: string | null;
};