# lib/api-zod/src/generated/types/appPatterns.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LogVision — Apache access log analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { HeatmapCell } from "./heatmapCell";
import type { HourlyTotal } from "./hourlyTotal";

export interface AppPatterns {
  appName: string;
  heatmap: HeatmapCell[];
  hourlyTotals: HourlyTotal[];
}