# lib/api-zod/src/generated/types/appSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LogVision — Apache access log analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface AppSummary {
  name: string;
  requestCount: number;
  uniqueIPs: number;
  peakHour: number;
  peakDayOfWeek: number;
  /** @nullable */
  topPage?: string | null;
}