# lib/api-zod/src/generated/types/heatmapCell.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LogVision — Apache access log analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface HeatmapCell {
  dayOfWeek: number;
  hour: number;
  requestCount: number;
}