# lib/api-zod/src/generated/types/logSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LogVision — Apache access log analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface LogSession {
  id: number;
  filename: string;
  label: string;
  totalLines: number;
  parsedLines: number;
  /** @nullable */
  dateFrom: string | null;
  /** @nullable */
  dateTo: string | null;
  createdAt: string;
}