# lib/api-zod/src/generated/types/pageStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LogVision — Apache access log analytics API
 * OpenAPI spec version: 0.1.0
 */
import type { HourlyTotal } from "./hourlyTotal";
import type { PageStatStatusCodes } from "./pageStatStatusCodes";

export interface PageStat {
  url: string;
  requestCount: number;
  uniqueIPs: number;
  /** @nullable */
  avgBytes?: number | null;
  statusCodes: PageStatStatusCodes;
  hourlyDistribution?: HourlyTotal[];
}