# lib/api-zod/src/generated/types/timelinePoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LogVision — Apache access log analytics API
 * OpenAPI spec version: 0.1.0
 */

export interface TimelinePoint {
  timestamp: string;
  requestCount: number;
  /** @nullable */
  uniqueIPs?: number | null;
}